Single Page Application and Dependencies
Merged the multiple .html files to make our mobile app SPA (Single Page Application), using a custom page Router. This allowed us to use animations on the transition of the pages that will lead to more native feel and better user experience.

Also, dependencies were changed to load locally and not from CDN, in order to be able to use the app offline.